Realmのクライアントを実装
